pyproject.toml
[tool.poetry]
name = "sanae_bot"
version = "0.1.0"
description = "A QQBot base on nonebot2"
authors = ["AB_Solution"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "2.0.0a10"
ujson = "4.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
